backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Config & validation
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

# LLM + Agent (OpenAI SDK talks to Groq via base_url override — no groq package needed)
openai==1.54.0

# Retrieval
tavily-python==0.5.0
chromadb==0.5.18
onnxruntime==1.19.2
pypdf==5.1.0
beautifulsoup4==4.12.3
lxml==5.3.0

# HTTP & async
httpx==0.27.2
aiofiles==24.1.0

# State / cache
redis==5.2.0

# Rate limiting
slowapi==0.1.9

# Logging
loguru==0.7.2
